# BTreeMap/BTreeSet::from_iter: use bulk building to improve the performance