@Mark-Simulacrum 2021-09-04
@bors try @rust-timer queue 

Generally this seems OK, though I am mildly wary of the added code complexity. I don't think it's anything too worrisome though and in practice sorting is not typically that complicated.