@pnkfelix 2021-09-08
As predicted, during weekly performance triage this was tagged as causing [mixed changes to rustc instruction counts](https://perf.rust-lang.org/compare.html?start=11bbb5231349a0a144d86d5c0c21061a06d1969d&end=ffaf857045f4f4d8bb563e0a5077f9b065f42916&stat=instructions:u).

Its already been appropriately labelled; I'm just adding the link above for ease of reference in future.