@tanriol 2021-09-17
Sounds like `from_iter`'s new memory behavior should at least be documented somewhere.